Fix signed/unsigned comparison warning
Part of ongoing Linux work.

+++ src/ltablib.c
@@ -283,7 +283,7 @@ static int wipe (lua_State *L) {
 static int tremovemulti (lua_State *L) {
   luaL_checktype(L, 1, LUA_TTABLE);
 
-  const size_t length = lua_objlen(L, 1);
+  const int length = (int) lua_objlen(L, 1);
   const int index = luaL_optint(L, 2, length);
   const int count = luaL_optint(L, 3, 1);
